requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-decouple==3.5
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
stripe==2.61.0
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
